fix(connect): improve missing browser message
